requirements.txt
openai>=1.0.0
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0
psutil>=5.9.0
Send2Trash>=1.8.0
